requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
azure-communication-identity==1.0.1
azure-communication-phonenumbers==1.0.0
azure-communication-callautomation==1.0.0
python-dotenv==0.19.1
gunicorn==20.1.0
werkzeug==2.0.1